Updated codes for the ESP32
Those are the functions that control the ESP32. They are integrated into the general routines in Touchscreen-Chamber-Gerion. Now that there are copies of those routines in "Preparation", "Parallel Visual discrimination", "Visual Discrimination",... all of those need to be updated, hence the new branch.
I think the ideal structure of this Repo should be a single function that realized the entire setup and task (using those individual general module functions) and is only called with different settings from the outside: [Static Image or Movie], [Detection, Discrimination, Staircase], [Auto-rewarded Yes/No], ...