Created the new general TouchscreenChamber class
I tried to generalize a lot of common routines between Phase0 and Phase1. But their behavior is pretty different. Before I generalize this more I would like to compare this further with Phase2 at least.

I also added a copyright note for the new code as the repository is public now. The "CC BY-NC-SA" sounded sensible to me, but we should discuss this.