Added a routine so the mouse has to initiate the session
the mouse now has to touch the sinewave stimulus and collect the reward for the session to start. The stimulus is the same as in the pretraining phase, this should hopefully help the mice get into the session more easily and prevents that the first trial is already running while the mouse is still entering the chamber.